refactor configdb client package
